change type of id in blog response (long - int)

+++ src/main/java/vn/vinhdeptrai/skincarebookingsystem/dto/response/BlogResponse.java
@@ -4,7 +4,6 @@ import lombok.*;
 import lombok.experimental.FieldDefaults;
 
 import java.time.LocalDateTime;
-import java.util.Set;
 
 @Getter
 @Setter
@@ -13,7 +12,7 @@ import java.util.Set;
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class BlogResponse {
-    Long id;
+    int id;
     String title;
     String content;
     String thumbnail;
